feat: Adding the redirect setup inside a seperate file.
