+++ second_try/set_up_child_data.c
@@ -6,7 +6,7 @@
 /*   By: akuburas <akuburas@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/22 11:58:42 by akuburas          #+#    #+#             */
-/*   Updated: 2024/04/22 16:05:35 by akuburas         ###   ########.fr       */
+/*   Updated: 2024/04/22 16:09:05 by akuburas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,31 +27,7 @@ int	count_processes(t_input_list *start)
 	}
 }
 
-void	setup_redirects(t_shelldata *data, int index)
-{
-	t_input_list	*temp;
 
-	temp = data->input_list;
-	while (index > 0)
-	{
-		while (temp->type != PIPE)
-			temp = temp->next;
-		temp = temp->next;
-		index--;
-	}
-	while (temp && temp->type != PIPE)
-	{
-		if (temp->type == REDIRECT_INPUT)
-			handle_redirect_input(&data->child_data[index], temp->next);
-		if (temp->type == REDIRECT_OUTPUT)
-			handle_redirect_output(&data->child_data[index], temp->next);
-		if (temp->type == REDIRECT_APPEND)
-			handle_redirect_append(&data->child_data[index], temp->next);
-		temp = temp->next;
-		if (data->child_data[index].exit_value != 0)
-			return ;
-	}
-}
 
 void	create_child_data(t_shelldata *data, int amount)
 {
